Objective-C version
Looking to revert to Objective-C for development due to the hassle of integrating
C libraries into Vala. There's a need to use libjpeg to read JPEG files. This
doesn't have a Vala binding at the moment and it seems its a pain to add one. The
other option is GtkPixBuf but I don't want to depend on Gtk. There's also the
hassle of trying to get Vala running on Windows.

This first commit was developed against GNUStep on Windows and it seems to work well.